+++ data.csv
@@ -1,11 +1,46 @@
-,Unnamed: 0,screenshots,bird_tf
-0,0.0,screenshots/screenshot_2024-03-18 19:25:42.png,True
-1,1.0,screenshots/screenshot_2024-03-18 19:25:51.png,True
-2,2.0,screenshots/screenshot_2024-03-18 19:26:01.png,True.
-3,3.0,screenshots/screenshot_2024-03-18 19:26:09.png,True
-4,4.0,screenshots/screenshot_2024-03-18 19:26:19.png,False
-0,,screenshots/screenshot_2024-03-18 19:32:22.png,False
-1,,screenshots/screenshot_2024-03-18 19:32:31.png,False.
-2,,screenshots/screenshot_2024-03-18 19:32:40.png,False.
-3,,screenshots/screenshot_2024-03-18 19:32:51.png,False.
-4,,screenshots/screenshot_2024-03-18 19:33:00.png,False
+,Unnamed: 0.2,Unnamed: 0.1,Unnamed: 0,screenshots,bird_tf
+0,0.0,0.0,0.0,screenshots/screenshot_2024-03-18 19:25:42.png,True
+1,1.0,1.0,1.0,screenshots/screenshot_2024-03-18 19:25:51.png,True
+2,2.0,2.0,2.0,screenshots/screenshot_2024-03-18 19:26:01.png,True.
+3,3.0,3.0,3.0,screenshots/screenshot_2024-03-18 19:26:09.png,True
+4,4.0,4.0,4.0,screenshots/screenshot_2024-03-18 19:26:19.png,False
+5,5.0,0.0,,screenshots/screenshot_2024-03-18 19:32:22.png,False
+6,6.0,1.0,,screenshots/screenshot_2024-03-18 19:32:31.png,False.
+7,7.0,2.0,,screenshots/screenshot_2024-03-18 19:32:40.png,False.
+8,8.0,3.0,,screenshots/screenshot_2024-03-18 19:32:51.png,False.
+9,9.0,4.0,,screenshots/screenshot_2024-03-18 19:33:00.png,False
+10,0.0,,,screenshots/screenshot_2024-03-25 14:41:07.png,False
+11,1.0,,,screenshots/screenshot_2024-03-25 14:41:13.png,False.
+12,2.0,,,screenshots/screenshot_2024-03-18 19:26:01.png,True
+13,3.0,,,screenshots/screenshot_2024-03-18 19:25:51.png,True
+14,4.0,,,screenshots/screenshot_2024-03-18 19:25:42.png,True
+15,5.0,,,screenshots/screenshot_2024-03-25 14:40:56.png,False
+16,6.0,,,screenshots/screenshot_2024-03-25 14:41:02.png,False
+17,7.0,,,screenshots/screenshot_2024-03-18 19:32:22.png,False
+18,8.0,,,screenshots/screenshot_2024-03-18 19:32:31.png,False
+19,9.0,,,screenshots/screenshot_2024-03-18 19:33:00.png,False
+20,10.0,,,screenshots/screenshot_2024-03-25 14:41:18.png,False
+21,11.0,,,screenshots/screenshot_2024-03-18 19:32:40.png,False
+22,12.0,,,screenshots/screenshot_2024-03-18 19:26:09.png,True
+23,13.0,,,screenshots/screenshot_2024-03-18 19:26:19.png,False
+24,14.0,,,screenshots/screenshot_2024-03-18 19:32:51.png,False
+0,,,,screenshots/screenshot_2024-03-25 14:41:07.png,False
+1,,,,screenshots/screenshot_2024-03-25 14:41:13.png,False
+2,,,,screenshots/screenshot_2024-03-18 19:26:01.png,True
+3,,,,screenshots/screenshot_2024-03-18 19:25:51.png,True
+4,,,,screenshots/screenshot_2024-03-18 19:25:42.png,True
+5,,,,screenshots/screenshot_2024-03-25 14:40:56.png,False
+6,,,,screenshots/screenshot_2024-03-25 14:41:02.png,False
+7,,,,screenshots/screenshot_2024-03-25 14:49:03.png,False
+8,,,,screenshots/screenshot_2024-03-25 14:48:57.png,False
+9,,,,screenshots/screenshot_2024-03-25 14:48:52.png,False
+10,,,,screenshots/screenshot_2024-03-25 14:48:46.png,False.
+11,,,,screenshots/screenshot_2024-03-25 14:49:08.png,False
+12,,,,screenshots/screenshot_2024-03-18 19:32:22.png,False.
+13,,,,screenshots/screenshot_2024-03-18 19:32:31.png,False
+14,,,,screenshots/screenshot_2024-03-18 19:33:00.png,False
+15,,,,screenshots/screenshot_2024-03-25 14:41:18.png,False.
+16,,,,screenshots/screenshot_2024-03-18 19:32:40.png,False
+17,,,,screenshots/screenshot_2024-03-18 19:26:09.png,True
+18,,,,screenshots/screenshot_2024-03-18 19:26:19.png,False
+19,,,,screenshots/screenshot_2024-03-18 19:32:51.png,False
