+++ data_new.csv
@@ -1,0 +1,46 @@
+screenshots,bird_tf
+screenshots/screenshot_2024-03-18 19:25:42.png,True
+screenshots/screenshot_2024-03-18 19:25:51.png,True
+screenshots/screenshot_2024-03-18 19:26:01.png,True.
+screenshots/screenshot_2024-03-18 19:26:09.png,True
+screenshots/screenshot_2024-03-18 19:26:19.png,False
+screenshots/screenshot_2024-03-18 19:32:22.png,False
+screenshots/screenshot_2024-03-18 19:32:31.png,False.
+screenshots/screenshot_2024-03-18 19:32:40.png,False.
+screenshots/screenshot_2024-03-18 19:32:51.png,False.
+screenshots/screenshot_2024-03-18 19:33:00.png,False
+screenshots/screenshot_2024-03-25 14:41:07.png,False
+screenshots/screenshot_2024-03-25 14:41:13.png,False.
+screenshots/screenshot_2024-03-18 19:26:01.png,True
+screenshots/screenshot_2024-03-18 19:25:51.png,True
+screenshots/screenshot_2024-03-18 19:25:42.png,True
+screenshots/screenshot_2024-03-25 14:40:56.png,False
+screenshots/screenshot_2024-03-25 14:41:02.png,False
+screenshots/screenshot_2024-03-18 19:32:22.png,False
+screenshots/screenshot_2024-03-18 19:32:31.png,False
+screenshots/screenshot_2024-03-18 19:33:00.png,False
+screenshots/screenshot_2024-03-25 14:41:18.png,False
+screenshots/screenshot_2024-03-18 19:32:40.png,False
+screenshots/screenshot_2024-03-18 19:26:09.png,True
+screenshots/screenshot_2024-03-18 19:26:19.png,False
+screenshots/screenshot_2024-03-18 19:32:51.png,False
+screenshots/screenshot_2024-03-25 14:41:07.png,False
+screenshots/screenshot_2024-03-25 14:41:13.png,False
+screenshots/screenshot_2024-03-18 19:26:01.png,True
+screenshots/screenshot_2024-03-18 19:25:51.png,True
+screenshots/screenshot_2024-03-18 19:25:42.png,True
+screenshots/screenshot_2024-03-25 14:40:56.png,False
+screenshots/screenshot_2024-03-25 14:41:02.png,False
+screenshots/screenshot_2024-03-25 14:49:03.png,False
+screenshots/screenshot_2024-03-25 14:48:57.png,False
+screenshots/screenshot_2024-03-25 14:48:52.png,False
+screenshots/screenshot_2024-03-25 14:48:46.png,False.
+screenshots/screenshot_2024-03-25 14:49:08.png,False
+screenshots/screenshot_2024-03-18 19:32:22.png,False.
+screenshots/screenshot_2024-03-18 19:32:31.png,False
+screenshots/screenshot_2024-03-18 19:33:00.png,False
+screenshots/screenshot_2024-03-25 14:41:18.png,False.
+screenshots/screenshot_2024-03-18 19:32:40.png,False
+screenshots/screenshot_2024-03-18 19:26:09.png,True
+screenshots/screenshot_2024-03-18 19:26:19.png,False
+screenshots/screenshot_2024-03-18 19:32:51.png,False
