+++ screenshot_script.py
@@ -5,92 +5,34 @@ from selenium.webdriver.common.keys import Keys
 import time
 from datetime import datetime
 import base64
-import requests
+
 import os
 from os import listdir
 import pandas as pd
 import os.path
 
-screenshot_names = []
-results = []
+def take_screenshot(cam_url, chromedriver_path, iterations_count, screenshots_frequency):
+  screenshot_names = []
+  # Specify Chrome Options for headless
+  options = webdriver.ChromeOptions()
+  # Headless mode
+  options.add_argument('--headless') 
+  options.add_argument('--disable-extensions')
+  options.add_argument('--disable-gpu')
+  options.add_argument('--no-sandbox') 
+  options.add_argument('--disable-dev-shm-usage')
+  # Autoplay videos
+  options.add_argument('--autoplay-policy=no-user-gesture-required')
 
-def encode_image(image_path):
-  with open(image_path, "rb") as image_file:
-    return base64.b64encode(image_file.read()).decode('utf-8')
+  # Set up the Chrome service and initialize the Chrome driver
+  service = Service(executable_path=chromedriver_path)
+  driver = webdriver.Chrome(service=Service(executable_path=chromedriver_path), options=options)
+  driver.get(cam_url)
 
-def process_image_and_prompt(input_image=None):
+  # Wait for the live video to load
+  time.sleep(5)
 
-    # Path to your image
-#     image_path = "path_to_your_image.jpg"
-
-    # Getting the base64 string
-    base64_image = encode_image(input_image)
-
-    headers = {
-      "Content-Type": "application/json",
-      "Authorization": f"Bearer {OPENAI_API_KEY}"
-    }
-
-    payload = {
-      "model": "gpt-4-vision-preview",
-      "messages": [
-        {
-          "role": "user",
-          "content": [
-            {
-              "type": "text",
-              "text": "Does this image contain a bird? Only respond with True or False."
-            },
-            {
-              "type": "image_url",
-              "image_url": {
-                "url": f"data:image/jpeg;base64,{base64_image}"
-              }
-            }
-          ]
-        }
-      ],
-      "max_tokens": 300
-    }
-
-    response = requests.post("https://api.openai.com/v1/chat/completions", headers=headers, json=payload)
-    if response.status_code == 200:
-        json_response = response.json()
-        content = json_response['choices'][0]['message']['content']
-        return content
-    else:
-        return f"Error: {response.status_code} - {response.reason}"
-
-# Specify the path to Chromedriver
-# Specify bird cam URL
-# Specify iterations
-# Specify screenshots frequency
-# Adjust if necessary 
-chromedriver_path = './chromedriver-mac-arm64/chromedriver'
-cam_url = 'https://www.youtube.com/watch?v=Mez4F1iFxxM&ab_channel=CornellLabBirdCams'
-iterations_count = 5
-screenshots_frequency = 5
-
-# Specify Chrome Options for headless
-options = webdriver.ChromeOptions()
-# Headless mode
-options.add_argument('--headless') 
-options.add_argument('--disable-extensions')
-options.add_argument('--disable-gpu')
-options.add_argument('--no-sandbox') 
-options.add_argument('--disable-dev-shm-usage')
- # Autoplay videos
-options.add_argument('--autoplay-policy=no-user-gesture-required')
-
-# Set up the Chrome service and initialize the Chrome driver
-service = Service(executable_path=chromedriver_path)
-driver = webdriver.Chrome(service=Service(executable_path=chromedriver_path), options=options)
-driver.get(cam_url)
-
-# Wait for the live video to load
-time.sleep(5)
-
-try:
+  try:
     '''
     # Will use this only with GUI (not headless)
 
@@ -112,21 +54,9 @@ try:
         driver.save_screenshot(img_name)
         time.sleep(screenshots_frequency)
         screenshot_names.append(img_name)
-        results.append(process_image_and_prompt(img_name))
+        # results.append(process_image_and_prompt(img_name))
     
-finally:
-    # Clean up and close the browser window
-    driver.quit()
-
-DIR = "data.csv"
-if os.path.isfile(DIR):
-    df = pd.read_csv(DIR)
-    df_new = pd.DataFrame({"screenshots" : screenshot_names, "bird_tf": results})
-    data = pd.concat([df, df_new])
-else:
-    data = pd.DataFrame({"screenshots" : screenshot_names, "bird_tf": results})
-
-# for img in os.listdir(DIR):
-#     screenshot_names.append(img)
-#     results.append(process_image_and_prompt(f"./screenshots/{img}"))
-data.to_csv("data.csv")
+  finally:
+      # Clean up and close the browser window
+      driver.quit()
+  return screenshot_names
